requirements.txt
Django==3.2.13
django-currentuser==0.5.3
django-filter==21.1
django-seed==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0